Adds 'hiddenUntilFinishLoad' param
This param is true by default, so it does not change the current behaviour: a hidden webview is automatically revealed once stateChanged.finishLoad is received.

It can be set to false so that a hidden webview is not automatically revealed when the page has finished loading.

This way, the user can keep the webview hidden as long as required (i.e. to wait for custom javascript evaluation after the page finishes loading).